21.11.2019 19:40

+++ Lift/Models/StartupConfigurations.cs
@@ -11,7 +11,7 @@ namespace Lift.Models
     {
         public bool ModelWorkSystem { get; set; } = false;
         public int Floors { get; set; } = 12;
-        public int Lifts { get; set; } = 3;
+        public decimal Lifts { get; set; } = 3;
         public int PeopleInLift { get; set; } = 5;
         public double SpeedOfLift { get; set; } = 1.5;
         public double AccelerationOfLift { get; set; } = 0.5;
